requirements.txt
Django==2.0.4  
Celery==4.1.0  
djangorestframework==3.8.2 
matplotlib==2.2.2
sympy==1.1.1 
nose==1.3.7
numpy==1.13.3
scipy==1.0.0
eyed3==0.8.5
simplejson==3.13.2
ipython==6.2.1
jupyter==1.0.0
pandas==0.20.3
psycopg2-binary==2.7.4  
redis==2.10.6  
firebase-admin==2.9.1
opencv-python==3.4.0.12
soundfile==0.10.1
PyWavelets==0.5.2

